dpa_offload: Use a limited number of statistics handling threads
If we are creating a new statistics worker thread for each new counters
request we risk allowing these threads to spawn uncontrollably in the
system. The number of worker threads was therefore limited to
MAX_NUM_OF_THREADS and each thread is capable now of acquiring work from
the work queue by itself. When no more work is available the thread
dies. It will be respawned by the DPA Stats driver once work becomes
available again.

Change-Id: Ia020ea559653d42d654f5781a844c7cb3d7045fb
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/4594
Reviewed-by: Zanoschi Aurelian-B43522 <Aurelian.Zanoschi@freescale.com>
Reviewed-by: Garg Vakul-B16394 <vakul@freescale.com>
Tested-by: Garg Vakul-B16394 <vakul@freescale.com>