simple_proto: fix PDCP test vector generation for performance mode
In simple proto the test vector for performance mode is generated
through increments of a uint8_t variable. In case the HFN override
mode is enabled for PDCP, the function responsible for creating
the test vector is overriden and is called per each frame. This
leads to all the frames having the same content (0x00, 0x01, ...)
while the checking is expecting a value that is always incremented.
This patch ensures that the correct values are generated even in
HFN override mode.

Change-Id: I3164500e590132c5667585c3dcbdfef61ca6b404
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/4737
Reviewed-by: Geanta Neag Horia Ioan-B05471 <horia.geanta@freescale.com>
Reviewed-by: Stoica Cristian-B18196 <cristian.stoica@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>