lib_crypto: fix free'ing of FQs
In order to free FQs, the crypto library maintained arrays for each
FQ that signaled if the FQ can be retired or not. The arrays were
completed on QMan dequeue callback. This patch aims to enable a
more robust behavior by using directly the appropriate QMan APIs
for checking if the FQs can be destroyed.

Change-Id: Ie3003ac1439709dfcd9fe452ba749017cc985bd8
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/4747
Reviewed-by: Geanta Neag Horia Ioan-B05471 <horia.geanta@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>