simple_proto, simple_crypto: add SIGINT signal handling
Both simple_proto and simple_crypto would bail out in case the
user pressed Ctrl+C at the console. In this case, the allocated
resources (FQs, etc.) are not free'ed properly, and if the
application is invoked again, they can no longer be acquired.
This patch aims to solve this problem by installing a SIGINT
handler on the main application thread which "cancels" the
worker threads.

Change-Id: Icfce2410f9dd7054d6809f405a450d6427fa0693
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/5013
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>