drivers/compat: Changed locks to use adaptive mutexes
QBman drivers require locks when portal sharing is enabled. These locks
are typically required for very small durations (e.g. copying enqueue
command in EQCR). These locks are acquired frequently. So contention for
the lock often occurs among portal sharing threads. In such a case, the
default mutex lock (which is optimized for non-contended case) causes a
futex operation and a context switch on the thread which finds lock to
be unavailable. Since locks are held for smaller duration, it is often
efficient for waiting threads to spin-wait for few cycles rather than
incur a context switch immediately.

Adaptive pthread mutexes provide the above mentioned lock behavior. For
4 threads sharing a portal, the hello reflector application shows double
the performance with adaptive mutexes compared to default mutexes.

Signed-off-by: Vakul Garg <vakul@freescale.com>
Change-Id: I9d9fe3555e761bf5cf5e8d40d68f16238e154ead
Reviewed-on: http://git.am.freescale.net:8181/5019
Reviewed-by: Thorpe Geoff-R01361 <Geoff.Thorpe@freescale.com>