usdpaa: move defines from internal headers to external headers
Some #define-s which are inside the internal headers of USDPAA
(include/internal/*.h) were used by the applications. Since its
forbidden to include internal headers into the applications
directly, these defines need to be moved to the external
headers (include/usdpaa/*.h). This will not pose a problem for
the users of the internal headers since all internal headers
include the external ones.

Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Change-Id: I29e660a98802a1062125401db591689e81675d7f
Reviewed-on: http://git.am.freescale.net:8181/5214
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>