apps: rename mutex functions
This patch renames mutex related functions in apps by
replacing spin_lock_* prefix with mutex_* prefix in order
to decrease the confusion of the user when using "spin_lock_*"
functions as mutex. The patch also moves back to the internal
compat.h header the spin-lock structures (which are needed
for compatibility with the Linux drivers).

Change-Id: I080cf4dba9a6afcf536975c343fb4ea93b2326fa
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/5503
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>