dpa_offload: Create detached DPA stats worker threads
The worker threads that are serving the user space requests are created
with default flags, meaning that they are "joinable". In this case, after
their death, the system waits for their parent to join them in in order to
release the resources consumed by the threads. This is not possible as the
parent thread is not controlled by the driver, but by the user application.
The solution to this is to create the DPA stats worker threads as "detached",
and when they die, the system automatically recycles their resources,
w/o waiting for any actions from their parent.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: Ibecc394c5ccd40bdf54300fcda93187f1e7fa503
Reviewed-on: http://git.am.freescale.net:8181/5504
Reviewed-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>