dpa_offload: Fix IPv4 options size field inconsistency
Due to a bad sync the ipv4_header defines the attribute options_size as
uint8_t, while the compat_ipv4_header defines it as unsigned int. The
compat_ipv4_header must define this attribute as uint8_t as well to
avoid problems when casting from one structure to the other.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I8933bfa8ef25789291874c279d18a70a2c142f6b
Reviewed-on: http://git.am.freescale.net:8181/5735
Reviewed-by: Bulie Radu-Andrei-B37577 <Radu.Bulie@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>