qman: Add enqueue api variant
Add a QMan enqueue api variant which invokes a user supplied callback
function just before writing the commit verb. The callback function
is invoked when the request is guaranteed to be enqueueud. The
standard enqueue request can return an error. Useful when the caller
is saving state information and isn't able to roll back the transaction.

Signed-off-by: Jeffrey Ladouceur <Jeffrey.Ladouceur@freescale.com>
Change-Id: Id5ebfd95e8ba6a64021480e7e9260996fb7ced65
Reviewed-on: http://git.am.freescale.net:8181/6041
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>