simple_crypto: fix incorrect usage of stack variables
Due to inattention to detail, the crypto_info variable on the
stack of the main function is not memset-ed to 0 before being
used, leading to erratic behavior through the application.
Also, the crypto_info variable is passed through value, as
opossed to pointer to the various init_rtv_* functions, which
locally set the authnct member of the crypto_info structure.
This leads to issues because for certain vectors, not setting
this variable leads to crashes of the application because
the shared descriptor is not properly initialized.

Change-Id: Ibcf73f999c198996ae565c7aa0b58e947a85c1d1
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/6300
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>