simple_proto: fix PDCP decap verification if HFN override is enabled
If HFN override is enabled, then due to the way the encap buffer is
created (for each buffer and each iteration, the content of one byte
in the buffer is incremented by 1, as opposed to being incremented
only per each buffer, as the "main" set_enc_buf(...) function does
in the main body of the simple_proto application). This leads to
errors when more than one iteration is used, because the
test_dec_match(...) function in the main body of simple_proto
application expects that after each iteration, each buffer
starts with 0x00, 0x01... etc.

Change-Id: I641bc5411a27c2aac55ed43d39a27fd1cc827440
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/19948
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>