dpa_offload: Avoid stats worker threads starvation
The DPA statistics module worker threads are sharing the asynchronous
requests pool with the application threads. All these threads are
locking the same resource. If they are many, the application threads may
starve the stats worker threads by keeping the requests pool locked too
long. This blocks the process as the worker threads are unable to recycle
the processed requests and the application threads are constantly trying
to acquire new requests.

The asynchronous requests pool and the user space requests list were
re-implemented as FIFO queues where the consumers and producers would not
wait for each other. The producers will need to synchronize between them
and the consumers (worker threads) the same. Now the producers do not need
to synchronize with consumers other than waiting for FIFO queue content.

Change-Id: Ibc33d6b1c959ac9d59966d63251321e94b14763f
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/6597
Reviewed-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>