ipsec_offload: Add aggregation option
When aggregation is enabled at application startup,
inbound SAs post decrytpion frame queues are configured to
point to the O/H port pre encryption on outbound direction.
Thus all decrypted traffic on inbound direction will be
redirected to outbound and in case of policy hit, it will be
encrypted. In  a more generic way, if there are the following
inbound SAs: SA1, SA2, SA3 and the decrypted traffic will match
outbound SA4 then the SA1 to SA3 on inbound will be aggregated
in SA4 on outbound.

Change-Id: Ia3428ddddebacb0c53df8a7a4616495ec7f06ba1
Signed-off-by: Radu Bulie <radu.bulie@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/8225
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>