drivers/of: fix check_compatible function
Under t2080 it was observed that the device tree node (represented by
struct dt_file) has 'len' equal to the length of string pointed by
'buf'. The existing assert() makes the function check_compatible
intolerant of such condition. Replaced assert with a simple 'if' check
to handle this condition in a tolerant way.

Signed-off-by: Vakul Garg <vakul@freescale.com>
Change-Id: I6de41ae6d541bc7ff5dc34db5cd3ee11285304a5
Reviewed-on: http://git.am.freescale.net:8181/7383
Reviewed-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>