classifier_demo: Add updated use case with CEETM for B4 platforms
The classifier demo use case had a facelift having now supproted both
IPv4 and IPv6 traffic flows. The classifier keys can be added
dinamically with add4/6 command or deleted with rem4/6. For the B4
platforms the application has now support for CEETM, for all AFxx
traffic classes and works in the same time on P4 platforms, but without
shaping and the other features provided by the Traffic Manager. Up to 64
flows are supported for both IPv4 and IPv6 traffic. The DSCP table that
is used for the shaping algorithm has 12 static entries each for one AF
traffic class. In case of P4 platforms all traffic is forwarded to the
first configured fqid of the TX port.
The statistics are split in two major categories, one for the
classification tables created and one for the traffic. On B4 the traffic
statistics contain the information about the CEETM CQ's and CGR's, while
on P4 platforms only the ethernet port statistics are printed.

Change-Id: I5d69e17d733013ee6a70bb0164723eba5a2fc649
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/8605
Reviewed-by: Nicolae-Sebastian Grigore <sebastian.grigore@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>