ppac: add SEC ERA retrieval support
Up to now, ppac library would assume the SEC ERA to be equal to 2.
This is not correct, since USDPAA is used on platforms on which
the SEC ERA ranges from 2 to 6. This patch adds the necessary code
and Makefile configuration needed for using the already existing
functions for retrieving and validating the SEC ERA.

Change-Id: I031ca069598598fd69e4f2d11cfffd665360b147
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/8630
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>