apps/ppac: fix SEC ERA validation on use cases not using SEC
Due to the way the SEC era mechanism is implemented, all apps
using ppac go through the process of validating the SEC era,
either retrieved from hardware or given by the user. This leads
to problems if the mechanism fails (because f.i. the SEC era
cannot be read from hardware and the user does not provide one
since the use case does not require SEC usage). This patch
solves the problem by implementing a callback that can be
invoked from the end-application which enables ppac to do
the SEC era retrieval and validation. Weak-linking is used
for implementing a default version of the function which
inhibits ppac from doing SEC related inquiries.

Change-Id: I6b6be7c1a7330d29097038b6ba372c38fc6d9731
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/8840
Reviewed-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>