simple_proto: fix PDCP decapsulation checking for HFN override
Change 3b043322e90be6c5553e0f0da3c9e14eca8c857e on USDPAA
did not take into account the fact that the test_dec_match_cb
function will also be called for non-ERA2 devices. The HFN
override mechanism is different for SEC ERA > 2 and as such
the checking is done starting with offset 4 in the output
buffer. This patch fixes the issue by allowing the correct
offset to be used (4 for SEC ERA = 2 and 0 for SEC ERA > 2)

Change-Id: I31a836731b88d3804a4aa8984a60f722a773e07e
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/9987
Reviewed-by: Cristian Stoica <cristian.stoica@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>