apps/lib: Add Jenkins hash function for hash table implementation
Add Jenkins hash function for hash table implementation in USDPAA. The
lookup function will hash a variable-length array into a 32-bit value.
Every bit of the key affects every bit of the return value. It will be
completed in 6 * len + 35 instruction.

The best hash table sizes are powers of 2, and is not suitable for
cryptographic purposes.

This implementation belongs to Bob Jenkins, bob_jenkins@burtleburtle.net
and is distributed under Public Domain. Source is
http://burtleburtle.net/bob/c/lookupa.c

Change-Id: I02c60ab8fbf6df895b9be3db65402effc96b902b
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/10420
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>