simple_proto: fix storage of descriptor buffer
The descriptor buffer address needs to be saved in the
protocol structure so that it can be free'ed at the end
of the test. Each core creates its own set of descriptors,
one per each FQ. The code did not take into account this
and thus it would store the pointer to the descriptor
which corresponds to the last FQ. This patch fixes this
issue by making the "descr" member of the protocol structure
an array with number of FQs elements.

Change-Id: Ie233eb22fb92c1b408eff0ac01e268dcf2d316d2
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/10955
Reviewed-by: Tudor-Dan Ambarus <tudor.ambarus@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>
(cherry picked from commit f3dc1e4ff12e5a634fb733a94570c0e5e881f157)
Reviewed-on: http://git.am.freescale.net:8181/11031