classifier_demo: Fix TX fqid read value for P4080DS
In case of P4080DS platform, due to lack of CEETM the application will
send the classified packages to the first tx fqid of configured port.
The tx fqid value was read at ppam_interface_init but every time the
value was 0. Now the value for the TX fqid is read only at first call of
ppam_interface_tx_fqid.

Change-Id: I38240a75815057c5f7a8d72beeb87638f705f50d
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/11400
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>