ipsec_offload: Offload policies/SAs even if VIF is down
Multiple IPSec intances required a mechanism of distributing
SAs/policies on diffrent instances. This mechanism is using the
IP set on the VIF interface. Therefore if multiple IPSec instances
are needed to be enabled VIF has to be up and IP should be set on it.

Stricly for testing ipsec_offload with the performance test,
I have enabled offloading of SAs/policies even if VIF is down. This
will fix the reported issue.

If VIFs are down and multiple instances of ipsec_offload are used then
the SA/policies will be offloaded on both FMANs.

Signed-off-by: Andrei Varvara <andrei.varvara@freescale.com>
Change-Id: I68cf4ffa44df96abdbb65351b85a5b38ad83805d
Reviewed-on: http://git.am.freescale.net:8181/11715
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>