dpa_offload: Fixes for 64bit KS/64 bit US mode
There is an issue when calling the IOCTL for init on DPA Stats caused by
parameters sent by value instead of reference. This is why they are
truncated in kernel-space.
Another issue is related to the t_Device fmlib internal data structure
that is out of sync with the one from DPA Stats implementation. This is
why the handle ID is not extracted correctly.

Change-Id: I2df138a219a8eb15604d40378bef31c83fa5b92e
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/11983
Reviewed-by: Anca Jeanina Floarea <anca.floarea@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>