ipsec_offload: Add fix for application quit
When quitting the application, SIGTERM signal
was sent to neigh and xfrm threads. SIGTERM
was caught by the main process and that caused
the termination of the main process without the
possibility to catch the sighal by the xfrm and
neigh threads. This led to memory leaks and errors
when restarting the application.
The patch fixes the above issue.

Change-Id: Iea5de2748a580f64dad825c53fd17bb7da56aab5
Signed-off-by: Radu Bulie <radu.bulie@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/12346
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>