dpa_offload: Fix for classifier_demo clean-up phase
At this moment the FMD driver is not capable of managing two or more
nodes that share another node with header manipulations configured. This
is for sure a big limitation, but it can be avoided by using a shared
node with prefilled entries and not performing any add/remove operations
on it.
The classifier_demo application was using two Exact Match tables for the
protocol classification based on IPSRC/IPDST/PROTO that shared another
Exact Match table containing the traffic rules based on DSCP. If the
DSCP table is populated by the application at runtime and header
manipulations are defined for the entries from the first tables, at
clean-up, even if all the remove operations are performed in the correct
order, at the point the pcd clean-up function is called the FMD will
return an error and will not allow to the application run correctly the
second time.
The DSCP table is defined statically in the policy file, the frame
queues are configured later, when a key is inserted in any of the
protocol classification tables.
This patch will avoid the limitation from FMD and permit the application
to run after a restart.

Change-Id: I354775534abdb53350872c3b0fb27b6808e9b385
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/12441
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Tested-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>