simple_crypto: update HMAC shared descriptor creation function
RTA allows the user to specify a certain subset
of the ICV value to be written in the output buffer.
This patch syncs the two software components (USDPAA & RTA),
by setting the truncation length to 0, which is translated
to "full ICV value output".
Also, the RTA function for HMAC shared descriptor no longer
takes a void pointer for the 4th parameter (the ICV),
just a value (<>0 if ICV checking is desired, or 0 otherwise).
This patch syncs FLIB and RTA on this change.

Change-Id: I42d01e808623721f453f43e73157eee81d5d05ec
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/14652
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Tested-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/22759