standalone-env: Set the correct "sysroot" parameter
The standalone-env building will be failed under the released
SDK 1.6 ISOs environment. But the building will be ok based on
the sdk-devel.git (super git) environment for 1.6.

The reason is the parameter of the "sysroot" for the compiling.

For example, to build the usdpaa applications for the
t4240qds-64b platform. After the setting of "export SDK_ROOT=..."
and "source standalone-env -m t4240qds-64b", the gcc will get
different "sysroot" values:

Under the sdk-devel.git (super git) environment for 1.6, the
"sysroot" will be right like this:

$ powerpc64-fsl-linux-gcc -print-sysroot
/home/freescale/sdk-devel-1.6/build_t4240qds-64b_release/tmp/
sysroots/t4240qds-64b

Under the released SDK 1.6 ISOs environment, the "sysroot"
will be error like this:

$ powerpc64-fsl-linux-gcc -print-sysroot
/local/workspace/yocto-iso-make/sdk-v1.6.x/b4860qds-64b/
build_b4860qds-64b_release/tmp/sysroots/b4860qds-64b

The error "sysroot" was defined based on the environment where
the SDK 1.6 released ISOs were built in.

So for the usdpaa standalone-env, the "sysroot" should be given
explicitly for the building under any SDK environment.

Signed-off-by: Liu Gang <Gang.Liu@freescale.com>
Change-Id: If3ec3412ae7f75bd77b8c302eaaab9791499326e
Reviewed-on: http://git.am.freescale.net:8181/14688
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/22501