dpa_offload: Improve memory consumption for user space statistics counters
The DPA statistics module was allocating a lot of memory at initialization
time trying to prepare for all possibilities of the users configuring
statistics. This allocation was moved from initialization to the point
where statistics counters are actually created. This allows the DPA stats
driver to know exactly the type of counters created by the user and optimize
memory consumption for those situations.

Change-Id: I417bbfa51e1ec45a363e9ed1e743c140f3e95575
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/19065
Reviewed-by: Anca Jeanina Floarea <anca.floarea@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/23192
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>