dpa_offload: Improve memory consumption for user space statistics counters
The DPA statistics module was allocating a lot of memory at initialization
time trying to prepare for all possibilities of the users configuring
statistics. This allocation was moved from initialization to the point
where statistics counters are actually created. This allows the DPA stats
driver to know exactly the type of counters created by the user and optimize
memory consumption for those situations.

Change-Id: I11c5ef9cd1bbddd42bfc61ef0958b43acee36e46
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/19068
Reviewed-by: Anca Jeanina Floarea <anca.floarea@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>