libfslcrypto: fix superfluous pthread_barrier_wait function calls
The main thread needs to wait for all the spawned threads to
finish initializing the local QMan thread before attempting
to move on. In the same manner, all threads need to wait for
the others to finish the aforementioned initialization.

While here, ensure that the return code from pthread_barrier_wait
is not EINVAL. As per the man page of pthread_barrier_wait:

"Upon successful completion, the pthread_barrier_wait() function
shall return PTHREAD_BARRIER_SERIAL_THREAD for a single (arbitrary)
thread synchronized at the barrier and zero for each of the other
threads. Otherwise, an error number shall be returned to indicate the error.

The pthread_barrier_wait() function may fail if:

EINVAL
    The value specified by barrier does not refer to an initialized barrier object.

This function shall not return an error code of [EINTR]. "

Change-Id: Ieea8a290b8f8dd126ae19a9e97a40e2a6c25df4e
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/19947
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Tested-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/22522
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>