standalone-env: fix SOC variable not exported
The ${SOC} variable, which depends on the target platform
is used in apps/Makefile.am, but it's not exported from
standalone-env. Thus it will not be defined and all apps
that depend on a SOC will not be compiled.
This patch exports the SOC variable so it can be used
throughout the Makefiles.

Change-Id: Icc33ef444adecd3a230f7488c740235556df7904
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/20913
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/22525