lib/mm: Relocate common library from apps
Since mm it is used in other external source code (e.g. NFAPI), as
well as usdpaa applications, this is the right place for it to stay.

In this case, the library was added in a separate directory from
drivers called 'lib' and the headers required were added in in a
separate include directory, namely 'include/lib'. Also, the library
header have a rule to be installed alongside the other headers.

Signed-off-by: Bogdan-Constantin Popescu <bogdan.c.popescu@freescale.com>
Change-Id: I5119c9bf0c47b8affc65fb2fa28bf5d3136a80c9
Reviewed-on: http://git.am.freescale.net:8181/31254
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Tested-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>