qbman/qman_high.c: Correct endianness for ARM
LNI commit rate was set incorrectly. The endianness of the ARM core is
not the same as the endianness of qman. Only excess rate traffic was
passing through, because the excess traffic rate set function
correctly converted endianness.  The issue was only observed on ARM
based architectures with DPAA1 datapath because PowerPC architectures
are big endian like qman so no conversion of endianness was needed

JIRA issue#: QUSDPA-850

Signed-off-by: Ahmed Mansour <Ahmed.Mansour@nxp.com>